# contracts/client-agent-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: Hermes DevOps Client Agent API
  version: 1.0.0
  description: |
    Runtime 对 Client Agent 的派单/取消/查询/设备/诊断接口。
    - 派单返回 202 仅表示"已入本地队列",执行进度经 callbacks-api 回传。
    - 幂等:重复 idempotency_key 返回已有任务状态,不重复执行。
    - 诊断只提供白名单探测,禁止任意 shell(红线 §14)。
security:
  - mtls: []
paths:
...
  /api/v1/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getTask
      summary: 查询本地任务状态
      responses:
        '200':
          description: 任务当前状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: cancelTask
      summary: 取消任务(尽力而为:kill 设备进程 + 清理)
      responses:
        '202':
          description: 取消已受理,最终以 task-events/results 回调的终态为准
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    mtls:
      type: mutualTLS
      description: Phase 3 强制双向 mTLS;Phase 1 允许自签部署,契约先行声明。
  responses:
...
    NotFound:
      description: 任务不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    TaskStatus:
      type: object
      required: [task_id, state, attempt]
      properties:
        task_id: { type: string }
        idempotency_key: { type: string }
        attempt: { type: integer, minimum: 1 }
        state:
          description: Client 本地生命周期状态(§9 status 的 Client 可见子集)
          enum:
            - QUEUED
            - ACCEPTED
            - PREPARING
            - DOWNLOADING
            - DEPLOYING
            - RUNNING
            - COLLECTING
            - COMPLETED
            - FAILED
            - TIMEOUT
            - CANCELED
        detail: { type: string }
        updated_at: { type: string, format: date-time, description: UTC 毫秒精度 }
...